rename generic-utils -> helm-control-scripts + init refactoring
