--- examples/generic-utils/src/init.rs
@@ -1,9 +1,0 @@
-use generic_utils::lib::*;
-
-fn main() {
-    let mut log_builder = Builder::from_default_env();
-    log_builder.target(Target::Stdout).filter(None, LevelFilter::Info).init();
-    info!("Utils version {}", env!("CARGO_PKG_VERSION"));
-
-    info!("This is mock up");
-}
