+++ helm-control-scripts/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "generic-utils"
+name = "helm-control-scripts"
 version = "0.1.0"
 authors = ["Aleksei Sizov <a.sizov@typeable.io>"]
 edition = "2018"
