+++ helm-control-scripts/src/create_update.rs
@@ -1,4 +1,4 @@
-use generic_utils::lib::*;
+use helm_control_scripts::lib::*;
 
 fn main() {
     let mut log_builder = Builder::from_default_env();
@@ -9,17 +9,31 @@ fn main() {
     let default_values: DefaultValues = serde_json::from_str(&envs.defaults).unwrap();
     let cli_opts = CliOpts::from_args();
     info!("Cli options received {:?}", &cli_opts);
+    let overrides = match overrides(&cli_opts) {
+        Some(inner) => inner,
+        None => vec![],
+    };
+    let domain_name = domain_name(&cli_opts);
+    info!("Domain generated for deployment: {}", &domain_name);
     let deployment_parameters = HelmDeploymentParameters::new(&cli_opts, &default_values, &envs);
+    let image_tag = match cli_opts.tag {
+        Some(tag) => tag,
+        None => {
+            error!("mandatory tag argument was not provided");
+            panic!();
+        }
+    };
+    helm_init(&envs, &deployment_parameters);
     let helm_cmd = HelmCmd {
         name: envs.helm_bin,
-        mode: HelmMode::Uninstall,
+        mode: HelmMode::UpgradeInstall,
         release_name: cli_opts.name,
-        release_domain: String::from(""),
+        release_domain: domain_name, 
         namespace: cli_opts.namespace,
         deployment_parameters: deployment_parameters,
-        overrides: vec![],
-        default_values: vec![],
-        image_tag: String::from("")
+        overrides: overrides,
+        default_values: default_values.default_overrides,
+        image_tag: image_tag,
     };
     info!("Generated Helm args: {:?}", &helm_cmd.args());
     match helm_cmd.run() {
